+++ CHANGELOG.md
@@ -54,6 +54,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Added capability to filter attributes per global ID - #4640 by @NyanKiyoshi.
 - Added capability to search product types by value (through the name) - #4647 by @NyanKiyoshi.
 - Add mutation to change the authenticated user's password - #4656 by @fowczarek
+- Add an functionality to sort products by their "minimal variant price" - #4416 by @derenio
 
 ## 2.8.0